Refactor manual vs managed single node clustering tests into separate packages.
