+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/cdi/ProtoStreamCDIFailoverTestCase.java
@@ -31,8 +31,8 @@ import org.jboss.as.test.clustering.cluster.cdi.webapp.IncrementorBean;
 import org.jboss.as.test.clustering.cluster.cdi.webapp.CDISerializationContextInitializer;
 import org.jboss.as.test.clustering.cluster.ejb.stateful.bean.Incrementor;
 import org.jboss.as.test.clustering.cluster.web.AbstractWebFailoverTestCase;
-import org.jboss.as.test.clustering.single.web.Mutable;
-import org.jboss.as.test.clustering.single.web.SimpleServlet;
+import org.jboss.as.test.clustering.managed.web.Mutable;
+import org.jboss.as.test.clustering.managed.web.SimpleServlet;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.EnterpriseArchive;
