+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/distributable-web_immutable_fine.xml
@@ -2,6 +2,6 @@
 <distributable-web xmlns="urn:jboss:distributable-web:1.0">
     <infinispan-session-management cache-container="web" cache="sso" granularity="ATTRIBUTE">
         <local-affinity/>
-        <immutable-class>org.jboss.as.test.clustering.single.web.Mutable</immutable-class>
+        <immutable-class>org.jboss.as.test.clustering.managed.web.Mutable</immutable-class>
     </infinispan-session-management>
 </distributable-web>
