+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/remote/FineTransactionalHotRodWebFailoverTestCase.java
@@ -27,8 +27,8 @@ import org.jboss.arquillian.container.test.api.TargetsContainer;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.test.clustering.cluster.web.AbstractWebFailoverTestCase;
-import org.jboss.as.test.clustering.single.web.Mutable;
-import org.jboss.as.test.clustering.single.web.SimpleServlet;
+import org.jboss.as.test.clustering.managed.web.Mutable;
+import org.jboss.as.test.clustering.managed.web.SimpleServlet;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
