+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/dispatcher/CommandDispatcherTestCase.java
@@ -19,21 +19,19 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.clustering.single.provider;
+package org.jboss.as.test.clustering.managed.dispatcher;
 
-import static org.jboss.as.test.clustering.cluster.AbstractClusteringTestCase.*;
-import static org.jboss.as.test.shared.integration.ejb.security.PermissionUtils.createPermissionsXmlAsset;
+import static org.jboss.as.test.clustering.cluster.AbstractClusteringTestCase.NODE_1;
 import static org.junit.Assert.*;
-
-import java.util.Collection;
-import java.util.PropertyPermission;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.as.test.clustering.cluster.provider.bean.ServiceProviderRetriever;
-import org.jboss.as.test.clustering.cluster.provider.bean.ServiceProviderRetrieverBean;
+import org.jboss.as.test.clustering.cluster.dispatcher.bean.ClusterTopology;
+import org.jboss.as.test.clustering.cluster.dispatcher.bean.ClusterTopologyRetriever;
+import org.jboss.as.test.clustering.cluster.dispatcher.bean.ClusterTopologyRetrieverBean;
 import org.jboss.as.test.clustering.ejb.EJBDirectory;
 import org.jboss.as.test.clustering.ejb.RemoteEJBDirectory;
+import org.jboss.as.test.shared.integration.ejb.security.PermissionUtils;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
@@ -41,29 +39,30 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
- * Validates that a service provider registration works in a non-clustered environment.
+ * Validates that a command dispatcher works in a non-clustered environment.
  * @author Paul Ferraro
  */
 @RunWith(Arquillian.class)
-public class ServiceProviderRegistrationTestCase {
-    private static final String MODULE_NAME = ServiceProviderRegistrationTestCase.class.getSimpleName();
+public class CommandDispatcherTestCase {
+    private static final String MODULE_NAME = CommandDispatcherTestCase.class.getSimpleName();
 
     @Deployment(testable = false)
     public static Archive<?> createDeployment() {
         WebArchive war = ShrinkWrap.create(WebArchive.class, MODULE_NAME + ".war");
-        war.addPackage(ServiceProviderRetriever.class.getPackage());
-        war.addAsManifestResource(createPermissionsXmlAsset(new PropertyPermission(NODE_NAME_PROPERTY, "read"), new RuntimePermission("getClassLoader")), "permissions.xml");
-        war.setWebXML(org.jboss.as.test.clustering.cluster.provider.ServiceProviderRegistrationTestCase.class.getPackage(), "web.xml");
+        war.addPackage(ClusterTopologyRetriever.class.getPackage());
+        war.addAsManifestResource(PermissionUtils.createPermissionsXmlAsset(new RuntimePermission("getClassLoader")), "permissions.xml");
+        war.setWebXML(org.jboss.as.test.clustering.cluster.dispatcher.CommandDispatcherTestCase.class.getPackage(), "web.xml");
         return war;
     }
 
     @Test
     public void test() throws Exception {
         try (EJBDirectory directory = new RemoteEJBDirectory(MODULE_NAME)) {
-            ServiceProviderRetriever bean = directory.lookupStateless(ServiceProviderRetrieverBean.class, ServiceProviderRetriever.class);
-            Collection<String> names = bean.getProviders();
-            assertEquals(1, names.size());
-            assertTrue(names.toString(), names.contains(NODE_1));
+            ClusterTopologyRetriever bean = directory.lookupStateless(ClusterTopologyRetrieverBean.class, ClusterTopologyRetriever.class);
+            ClusterTopology topology = bean.getClusterTopology();
+            assertEquals(1, topology.getNodes().size());
+            assertTrue(topology.getNodes().toString(), topology.getNodes().contains(NODE_1));
+            assertTrue(topology.getRemoteNodes().toString() + " should be empty", topology.getRemoteNodes().isEmpty());
         }
     }
 }
