+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/registry/RegistryTestCase.java
@@ -19,19 +19,21 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.clustering.single.dispatcher;
+package org.jboss.as.test.clustering.managed.registry;
 
-import static org.jboss.as.test.clustering.cluster.AbstractClusteringTestCase.NODE_1;
+import static org.jboss.as.test.clustering.cluster.AbstractClusteringTestCase.*;
+import static org.jboss.as.test.shared.integration.ejb.security.PermissionUtils.createPermissionsXmlAsset;
 import static org.junit.Assert.*;
+
+import java.util.Collection;
+import java.util.PropertyPermission;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.as.test.clustering.cluster.dispatcher.bean.ClusterTopology;
-import org.jboss.as.test.clustering.cluster.dispatcher.bean.ClusterTopologyRetriever;
-import org.jboss.as.test.clustering.cluster.dispatcher.bean.ClusterTopologyRetrieverBean;
+import org.jboss.as.test.clustering.cluster.registry.bean.RegistryRetriever;
+import org.jboss.as.test.clustering.cluster.registry.bean.RegistryRetrieverBean;
 import org.jboss.as.test.clustering.ejb.EJBDirectory;
 import org.jboss.as.test.clustering.ejb.RemoteEJBDirectory;
-import org.jboss.as.test.shared.integration.ejb.security.PermissionUtils;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
@@ -39,30 +41,29 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
- * Validates that a command dispatcher works in a non-clustered environment.
+ * Validates that a registry works in a non-clustered environment.
  * @author Paul Ferraro
  */
 @RunWith(Arquillian.class)
-public class CommandDispatcherTestCase {
-    private static final String MODULE_NAME = CommandDispatcherTestCase.class.getSimpleName();
+public class RegistryTestCase {
+    private static final String MODULE_NAME = RegistryTestCase.class.getSimpleName();
 
     @Deployment(testable = false)
     public static Archive<?> createDeployment() {
         WebArchive war = ShrinkWrap.create(WebArchive.class, MODULE_NAME + ".war");
-        war.addPackage(ClusterTopologyRetriever.class.getPackage());
-        war.addAsManifestResource(PermissionUtils.createPermissionsXmlAsset(new RuntimePermission("getClassLoader")), "permissions.xml");
-        war.setWebXML(org.jboss.as.test.clustering.cluster.dispatcher.CommandDispatcherTestCase.class.getPackage(), "web.xml");
+        war.addPackage(RegistryRetriever.class.getPackage());
+        war.addAsManifestResource(createPermissionsXmlAsset(new PropertyPermission(NODE_NAME_PROPERTY, "read"), new RuntimePermission("getClassLoader")), "permissions.xml");
+        war.setWebXML(org.jboss.as.test.clustering.cluster.registry.RegistryTestCase.class.getPackage(), "web.xml");
         return war;
     }
 
     @Test
     public void test() throws Exception {
-        try (EJBDirectory directory = new RemoteEJBDirectory(MODULE_NAME)) {
-            ClusterTopologyRetriever bean = directory.lookupStateless(ClusterTopologyRetrieverBean.class, ClusterTopologyRetriever.class);
-            ClusterTopology topology = bean.getClusterTopology();
-            assertEquals(1, topology.getNodes().size());
-            assertTrue(topology.getNodes().toString(), topology.getNodes().contains(NODE_1));
-            assertTrue(topology.getRemoteNodes().toString() + " should be empty", topology.getRemoteNodes().isEmpty());
+        try (EJBDirectory context = new RemoteEJBDirectory(MODULE_NAME)) {
+            RegistryRetriever bean = context.lookupStateless(RegistryRetrieverBean.class, RegistryRetriever.class);
+            Collection<String> names = bean.getNodes();
+            assertEquals(1, names.size());
+            assertTrue(names.toString(), names.contains(NODE_1));
         }
     }
 }
