+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/singleton/SingletonServiceTestCase.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.clustering.single.singleton;
+package org.jboss.as.test.clustering.managed.singleton;
 
 import static org.jboss.as.test.clustering.cluster.AbstractClusteringTestCase.NODE_1;
