+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/web/passivation/LocalCoarseSessionPassivationTestCase.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.test.clustering.single.web.passivation;
+package org.jboss.as.test.clustering.managed.web.passivation;
 
 import static org.jboss.as.test.clustering.cluster.AbstractClusteringTestCase.*;
 
@@ -30,16 +30,16 @@ import org.jboss.shrinkwrap.api.Archive;
 import org.junit.runner.RunWith;
 
 /**
- * Validates the correctness of session passivation events for a distributed session manager using a local, passivating cache and ATTRIBUTE granularity.
+ * Validates the correctness of session passivation events for a distributed session manager using a local, passivating cache and SESSION granularity.
  * @author Paul Ferraro
  */
 @RunWith(Arquillian.class)
-public class LocalFineSessionPassivationTestCase extends LocalSessionPassivationTestCase {
+public class LocalCoarseSessionPassivationTestCase extends LocalSessionPassivationTestCase {
 
-    private static final String MODULE_NAME = LocalFineSessionPassivationTestCase.class.getSimpleName();
+    private static final String MODULE_NAME = LocalCoarseSessionPassivationTestCase.class.getSimpleName();
 
     @Deployment(name = DEPLOYMENT_1, testable = false)
     public static Archive<?> deployment() {
-        return getBaseDeployment(MODULE_NAME).addAsWebInfResource(LocalSessionPassivationTestCase.class.getPackage(), "distributable-web-fine.xml", "distributable-web.xml");
+        return getBaseDeployment(MODULE_NAME).addAsWebInfResource(LocalSessionPassivationTestCase.class.getPackage(), "distributable-web-coarse.xml", "distributable-web.xml");
     }
 }
