+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/web/passivation/distributable-web-coarse.xml
