+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/web/passivation/jboss-web.xml
