+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/web/shared/jboss-all.xml
