+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/managed/web/web.xml
